[MIG] rotordc_sale_related_so: migration to 16.0
