+++ rotordc_sale_related_so/README.rst
@@ -17,7 +17,7 @@ RotorDC Sale Linked SO
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-coopiteasy%2Fcie--custom-lightgray.png?logo=github
-    :target: https://github.com/coopiteasy/cie-custom/tree/12.0/rotordc_sale_related_so
+    :target: https://github.com/coopiteasy/cie-custom/tree/16.0/rotordc_sale_related_so
     :alt: coopiteasy/cie-custom
 
 |badge1| |badge2| |badge3|
@@ -39,7 +39,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/coopiteasy/cie-custom/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/coopiteasy/cie-custom/issues/new?body=module:%20rotordc_sale_related_so%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/coopiteasy/cie-custom/issues/new?body=module:%20rotordc_sale_related_so%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -75,6 +75,6 @@ Current maintainer:
 
 |maintainer-robinkeunen| 
 
-This module is part of the `coopiteasy/cie-custom <https://github.com/coopiteasy/cie-custom/tree/12.0/rotordc_sale_related_so>`_ project on GitHub.
+This module is part of the `coopiteasy/cie-custom <https://github.com/coopiteasy/cie-custom/tree/16.0/rotordc_sale_related_so>`_ project on GitHub.
 
 You are welcome to contribute.
