+++ rotordc_sale_related_so/__manifest__.py
@@ -4,26 +4,21 @@
 
 {
     "name": "RotorDC Sale Linked SO",
-    "summary": """
-        Adds M2M links between SOs.""",
-    "version": "12.0.1.0.0",
-    "category": "Warehouse",
+    "summary": "Add M2M links between SOs",
+    "version": "16.0.1.0.0",
+    "category": "Sales/Sales",
     "website": "https://coopiteasy.be",
     "author": "Coop IT Easy SC",
     "maintainers": ["robinkeunen"],
     "license": "AGPL-3",
-    "application": False,
     "depends": [
         "sale_stock",
         "rotordc_report_picking_huge_sale_order",
     ],
-    "excludes": [],
     "data": [
         "views/sale_order_views.xml",
         "views/stock_picking_views.xml",
         "report/report_stockpicking_operations.xml",
         "security/ir.model.access.csv",
     ],
-    "demo": [],
-    "qweb": [],
 }
